Configure auth for frontend setup
Authentication is now made using either a frontend login form or via a HTTP request to the login endpoint /v1/accounts/login. The built-in login page is no longer used. Also added a logout endpoint as this was missing.

Lastly, updated allowed origins to include the hostname of the hosted instance of the frontend.

Co-Authored-By: Jakob Edvardsson <65956084+JakobEdvardsson@users.noreply.github.com>